CassandraEngineFactory for easy creation of an engine backed by a cassandra storage

+++ projects/copper-cassandra/cassandra-storage/src/test/java/org/copperengine/core/persistent/cassandra/CassandraTest.java
@@ -28,18 +28,21 @@ import org.junit.Test;
 
 public class CassandraTest extends CassandraUnitTest {
 
-    private static final CassandraEngineFactory factory = new CassandraEngineFactory();
+    private static UnitTestCassandraEngineFactory factory;
 
     @BeforeClass
     public static void setUpBeforeClass() throws Exception {
-        if (!skipTests())
-            factory.createEngine(true);
+        if (!skipTests()) {
+            factory = new UnitTestCassandraEngineFactory(true);
+            factory.getEngine().startup();
+        }
     }
 
     @AfterClass
     public static void tearDownAfterClass() throws Exception {
-        if (!skipTests())
+        if (!skipTests() && factory != null) {
             factory.destroyEngine();
+        }
     }
 
     @Test
@@ -48,14 +51,14 @@ public class CassandraTest extends CassandraUnitTest {
 
         List<String> cids = new ArrayList<>();
         for (int i = 0; i < 50; i++) {
-            final String cid = factory.engine.createUUID();
+            final String cid = factory.getEngine().createUUID();
             final TestData data = new TestData(cid, "foo");
             final WorkflowInstanceDescr<TestData> wfid = new WorkflowInstanceDescr<TestData>("org.copperengine.core.persistent.cassandra.workflows.TestWorkflow", data, cid, 1, null);
-            factory.engine.run(wfid);
+            factory.getEngine().run(wfid);
             cids.add(cid);
         }
         for (String cid : cids) {
-            Object response = factory.backchannel.wait(cid, 10000, TimeUnit.MILLISECONDS);
+            Object response = factory.backchannel.get().wait(cid, 10000, TimeUnit.MILLISECONDS);
             org.junit.Assert.assertNotNull("no response for workflow instance " + cid, response);
             org.junit.Assert.assertEquals("OK", response);
         }
@@ -66,11 +69,11 @@ public class CassandraTest extends CassandraUnitTest {
         assumeFalse(skipTests());
 
         for (int i = 0; i < 3; i++) {
-            final String cid = factory.engine.createUUID();
+            final String cid = factory.getEngine().createUUID();
             final TestData data = new TestData(cid, "foo");
             final WorkflowInstanceDescr<TestData> wfid = new WorkflowInstanceDescr<TestData>("org.copperengine.core.persistent.cassandra.workflows.TestWorkflow", data, cid, 1, null);
-            factory.engine.run(wfid);
-            Object response = factory.backchannel.wait(cid, 10000, TimeUnit.MILLISECONDS);
+            factory.getEngine().run(wfid);
+            Object response = factory.backchannel.get().wait(cid, 10000, TimeUnit.MILLISECONDS);
             org.junit.Assert.assertNotNull(response);
             org.junit.Assert.assertEquals("OK", response);
         }
